Minor fixes

+++ RLP_API/Models/RocketParameters.cs
@@ -6,7 +6,7 @@ namespace RLP_API.Models
         public double? LaunchMass { get; set; } = 0;
         public double? RocketLength { get; set; } = 0;
         public double? RocketDiameter { get; set; } = 0;
-        public int? SuccessfulRocketLaunches { get; set; } = 0;
-        public int? FailedRocketLaunches { get; set; } = 0;
+        public double? SuccessfulRocketLaunches { get; set; } = 0;
+        public double? FailedRocketLaunches { get; set; } = 0;
     }
 }
